Grid awareness and game over
The pieces are now able to tell when there are pieces under them. the x and y coordinates will update and tell which rows are complete, delete each line piece in the row, resulting in the row being deleted. The pieces will no longer fall through eachother also due to the x and y coordinates. The game over screen is implemented.